Minor modification to mounting hole
The mounting hole now is a real long hole instead of a square pocket. First
of course a square pocket cannot be manufactured exactly due to the mill form
and second this will leave way more material in vicinity of the lower edge
so the part won't be torn appart as easy while milling.